requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-genai
sentence-transformers
numpy
qdrant-client
pypdf
requests
python-multipart
PyPDF2
pydantic